# perf(root): reveal proofs separately from state updates