Unbundle sparse trie proof revealing from state updates, so that we can reveal proofs out-of-order and make the last reveal of the slowest proof as fast as possible. State updates for the sparse trie should still be applied in the order of sequence numbers, because updates can depend on each other.

This PR changes the `ProofSequencer` structure into the `StateUpdateSequencer`, and reveals proofs as soon as they come, only sequencing the state updates that should be applied in order.

I don't have a particularly good block on mainnet that will show the improvement, so here's just last 10 mainnet blocks. There's basically no difference, but in theory it should improve on situations when a slow proof is delaying not just all state updates, but also revealing other proofs.

| Mainnet Block | `newPayload` latency @ [main](https://github.com/paradigmxyz/reth/commit/590b58f) | `newPayload` latency @ [this branch](https://github.com/paradigmxyz/reth/pull/14077/commits/1e5ff3585e378eba2671593b613b6c74c9038f0a) |
|--------------|-----------------------------|------------------------------|
| 21744021     | 826.29157ms                 | 813.968945ms                 |
| 21744022     | 89.802651ms                 | 72.406547ms                  |
| 21744023     | 100.292771ms                | 89.415594ms                  |
| 21744024     | 59.863341ms                 | 45.230231ms                  |
| 21744025     | 148.03544ms                 | 151.032709ms                 |
| 21744026     | 30.603106ms                 | 33.761105ms                  |
| 21744027     | 308.403489ms                | 320.999399ms                 |
| 21744028     | 117.503467ms                | 128.529333ms                 |
| 21744029     | 72.660279ms                 | 66.99271ms                   |
| 21744030     | 54.977164ms                 | 51.821672ms                  |