@shekhirin 2025-01-31
This does not significantly improve the latency, because revealing proofs is already very fast. We can try applying both proofs and state updates out-of-order in another PR.